QueryDSL 도입 / draft, meeting부문 테스트까지 완료, 현재 schedule 테스트 케이스 정해두었고, 일정 등록/수정/취소에 대한 도메인 규칙 정정함. 이에 따라서 프로덕트 로직 정정하고 테스트 진행

+++ back/src/main/java/com/haruon/groupware/application/franchise/service/dto/DailySalesRequest.java
@@ -14,7 +14,7 @@ public record DailySalesRequest(
 
         LocalDate salesDate,
 
-        Double salesAmount,
+        Long salesAmount,
 
         Long orderCount
 ) {
