+++ back/src/test/java/com/haruon/groupware/domain/franchise/FranchiseDailySalesTest.java
@@ -20,7 +20,7 @@ class FranchiseDailySalesTest {
     void createSales_success() {
         String externalId = "external";
         LocalDate salesDate = LocalDate.of(2026,4,1);
-        Double salesAmount = 1000000.0;
+        Long salesAmount = 1000000L;
         Long orderCount = 1000L;
         Franchise franchise = getFranchise();
 
@@ -41,9 +41,9 @@ class FranchiseDailySalesTest {
 
         return Stream.of(
                 Arguments.of("매출액은 0 이상",
-                        -1.0, 1000L
+                        -1L, 1000L
                 ), Arguments.of("매출건수는 0이상",
-                        1000000.0, -1L
+                        1000000L, -1L
                 )
         );
     }
@@ -51,7 +51,7 @@ class FranchiseDailySalesTest {
     @ParameterizedTest(name = "{index} ==> {0}")
     @MethodSource("createSales_fail_params")
     @DisplayName("일매출건수 생성실패")
-    void createSales_fail_cases(String description, Double salesAmount, Long orderCount) {
+    void createSales_fail_cases(String description, Long salesAmount, Long orderCount) {
         assertThrows(IllegalStateException.class, () ->
                 FranchiseDailySales.create("external", LocalDate.now(), salesAmount, orderCount, getFranchise())
         );
@@ -61,13 +61,13 @@ class FranchiseDailySalesTest {
     @DisplayName("일매출건수 교체 테스트")
     void replace_sales_success() {
         LocalDate newSalesDate = LocalDate.of(2026,4,5);
-        Double newSalesAmount = 2000000.0;
+        Long newSalesAmount = 2000000L;
         Long newOrderCount = 2000L;
 
         FranchiseDailySales sales = FranchiseDailySales.create(
                 "external",
                 LocalDate.of(2026,4,1),
-                1000000.0,
+                1000000L,
                 1000L,
                 getFranchise());
 
@@ -86,9 +86,9 @@ class FranchiseDailySalesTest {
 
         return Stream.of(
                 Arguments.of("매출액은 0 이상",
-                        -1.0, 1000L
+                        -1L, 1000L
                 ), Arguments.of("매출건수는 0이상",
-                        1000000.0, -1L
+                        1000000L, -1L
                 )
         );
     }
@@ -96,11 +96,11 @@ class FranchiseDailySalesTest {
     @ParameterizedTest(name = "{index} ==> {0}")
     @MethodSource("replaceSales_fail_params")
     @DisplayName("일매출건수 교체실패")
-    void replaceSales_fail_cases(String description, Double salesAmount, Long orderCount) {
+    void replaceSales_fail_cases(String description, Long salesAmount, Long orderCount) {
         FranchiseDailySales sales = FranchiseDailySales.create(
                 "external",
                 LocalDate.of(2026,4,1),
-                1000000.0,
+                1000000L,
                 1000L,
                 getFranchise()
         );
